Move trace code to a separate file.
